src: reduce cpu profiler overhead
Reduce the overhead of the CPU profiler by suppressing SIGPROF signals
when sleeping / polling for events.  Avoids unnecessary wakeups when
the CPU profiler is active.

The benefits are quite substantial on systems where it is expensive to
query the system time (most virtualized environments).  Before, it was
common to see over 20% of wall clock time being spent on clock_gettime()
system calls.