lib: add internal binary heap implementation
Add a binary heap a.k.a. a priority queue, implemented as a binary tree.

Implementing it using an array is easier but JS arrays start to perform
poorly after they reach a certain size (about 2^17 elements.)

The current implementation doesn't use ES6 symbols due to performance
issues (symbol property lookups are currently about 5-10x slower when
compared to named proprty lookups) but that can probably be revisited
once we upgrade to V8 4.2 or 4.3.