vm: run runInThisContext() in last entered context
Change `vm.runInThisContext()` to use the last entered context, not the
top-level context that `vm.runInThisContext()` belongs to.

Note the call chain:

    1. `node::ContextifyScript::RunInThisContext()`
    2. `node::ContextifyScript::EvalMachine()`
    3. `v8::Script::BindToCurrentContext()`

The current context in `RunInThisContext()` is that of the top-level
context which was then bound to the script to execute.

It works when `vm.runInThisContext()` is called from the top-level
context (the common case and hence why this bug went unnoticed for
so long) but not when called from an inferior context.

As a pleasant side effect, this commit fixes a bug in the REPL where
writes to `process.stdout` and `process.stderr` still went to the real
stdout and stderr instead of being captured redirected.

Fixes previously failing test `known_issues/test-repl-require-context`
which has been moved to `test/parallel` and renamed to avoid a conflict
with an existing test of the same name.

Note that this commit changes the behavior of the following snippet:

    const s = '(function() { return vm.runInThisContext(this) })()';
    const f = vm.runInNewContext(s, vm);
    f();

Before this commit, it returned the |this| of the new context, now
it returns the |this| of the enclosing context.  Two steps forward,
one step back.

Fixes: https://github.com/nodejs/node/issues/7788
Refs: https://github.com/nodejs/node/issues/14757